+++ docs/_source/drl.agents.integration.rst
@@ -1,0 +1,21 @@
+drl.agents.integration package
+==============================
+
+Submodules
+----------
+
+drl.agents.integration.agent module
+-----------------------------------
+
+.. automodule:: drl.agents.integration.agent
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: drl.agents.integration
+   :members:
+   :undoc-members:
+   :show-inheritance:
