+++ docs/_source/drl.envs.wrappers.rst
@@ -10,17 +10,6 @@ Subpackages
    drl.envs.wrappers.stateful
    drl.envs.wrappers.stateless
 
-Submodules
-----------
-
-drl.envs.wrappers.integration module
-------------------------------------
-
-.. automodule:: drl.envs.wrappers.integration
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 Module contents
 ---------------
