--- docs/modules.rst
@@ -1,9 +1,0 @@
-pytorch_drl
-===========
-
-.. toctree::
-   :maxdepth: 4
-
-   drl
-   script
-   setup
