--- docs/script.rst
@@ -1,7 +1,0 @@
-script module
-=============
-
-.. automodule:: script
-   :members:
-   :undoc-members:
-   :show-inheritance:
