--- docs/setup.rst
@@ -1,7 +1,0 @@
-setup module
-============
-
-.. automodule:: setup
-   :members:
-   :undoc-members:
-   :show-inheritance:
