+++ drl/agents/integration/__init__.py
@@ -1,0 +1,5 @@
+from drl.agents.integration.agent import Agent
+
+__all__ = [
+    "Agent"
+]
