--- drl/envs/wrappers/integration.py
@@ -1,14 +1,0 @@
-from typing import Dict, Any
-import importlib
-
-
-def get_wrapper(env, cls_name, cls_args):
-    module = importlib.import_module('drl.envs.wrappers')
-    cls = getattr(module, cls_name)
-    return cls(env, **cls_args)
-
-
-def get_wrappers(env, **wrappers_spec: Dict[str, Dict[str, Any]]):
-    for cls_name, cls_args in wrappers_spec.items():
-        env = get_wrapper(env, cls_name, cls_args)
-    return env
